# docs/openapi.yml
openapi: 3.0.3
info:
  title: Team Dev Frontend Server API
  description: |-
  version: 0.2.2
servers:
  - url: https://team-dev-server-c10.fly.dev/api-docs/#/
...
paths:
...
  /units:
    post:
      tags:
        - unit
      summary: Create/Add Unit
      description: Create and add a unit to a module.
      operationId: CreateUnit
      requestBody:
        description: The name of the unit and the module Id. The module Id corresponds to the desired module that the unit is to be added to.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUnit'
      responses:
        '201':
          description: success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedUnit'
        '400':
          description: fail
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Error'
        '404':
          description: fail
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Error'
        '409':
          description: fail
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Error'
        '500':
          description: fail
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreatedUnit:
      type: object
      properties:
        status:
          type: string
          example: success
        data:
          properties:
            unit:
              properties:
                id:
                  type: integer
                name:
                  type: string
                createdAt:
                  type: string
                updatedAt:
                  type: string
...